+++ pkg/execution/state/state.go
@@ -38,8 +38,11 @@ var (
 	ErrPauseNotFound = fmt.Errorf("pause not found")
 	// ErrPauseLeased is returned when attempting to lease a pause that is
 	// already leased by another event.
-	ErrPauseLeased      = fmt.Errorf("pause already leased")
-	ErrIdentifierExists = fmt.Errorf("identifier already exists")
+	ErrPauseLeased       = fmt.Errorf("pause already leased")
+	ErrIdentifierExists  = fmt.Errorf("identifier already exists")
+	ErrFunctionCancelled = fmt.Errorf("function cancelled")
+	ErrFunctionComplete  = fmt.Errorf("function completed")
+	ErrFunctionFailed    = fmt.Errorf("function failed")
 )
 
 // RunStatus indicates the status for an individual function run.
@@ -64,63 +67,6 @@ func (i Identifier) IdempotencyKey() string {
 		key = i.RunID.String()
 	}
 	return fmt.Sprintf("%s:%s", i.WorkflowID, key)
-}
-
-// Pause allows steps of a function to be paused until some time in the future.
-// It pauses a specific workflow run via an Identifier, at a specific step in
-// the function as specified by Target.
-type Pause struct {
-	ID uuid.UUID `json:"id"`
-	// Identifier is the specific workflow run to resume.  This is required.
-	Identifier Identifier `json:"identifier"`
-	// Outgoing is the parent step for the pause.
-	Outgoing string `json:"outgoing"`
-	// Incoming is the step to run after the pause completes.
-	Incoming string `json:"incoming"`
-	// Expires is a time at which the pause can no longer be resumed.  This
-	// gives each pause of a function a TTL.  This is required.
-	//
-	// NOTE: the pause should remain within the backing state store for
-	// some perioud after the expiry time for checking timeout branches:
-	//
-	// If this pause has its OnTimeout flag set to true, we only traverse
-	// the edge if the event *has not* been received.  In order to check
-	// this, we enqueue a job that executes on the pause timeout:  if the
-	// pause has not yet been consumed we can safely assume the event was
-	// not received.  Therefore, we must be able to load the pause for some
-	// time after timeout.
-	Expires Time `json:"expires"`
-	// Event is an optional event that can resume the pause automatically,
-	// often paired with an expression.
-	Event *string `json:"event"`
-	// Expression is an optional expression that must match for the pause
-	// to be resumed.
-	Expression *string `json:"expression"`
-	// ExpressionData _optionally_ stores only the data that we need to evaluate
-	// the expression from the event.  This allows us to load pauses from the
-	// state store without round trips to fetch the entire function state.  If
-	// this is empty and the pause contains an expression, function state will
-	// be loaded from the store.
-	ExpressionData map[string]any `json:"data"`
-	// OnTimeout indicates that this incoming edge should only be ran
-	// when the pause times out, if set to true.
-	OnTimeout bool `json:"onTimeout"`
-	// DataKey is the name of the step to use when adding data to the function
-	// run's state after consuming this step.
-	//
-	// This allows us to create arbitrary "step" names for storing async event
-	// data from matching events in async edges, eg. `waitForEvent`.
-	//
-	// If DataKey is empty and data is provided when consuming a pause, no
-	// data will be saved in the function state.
-	DataKey string `json:"dataKey,omitempty"`
-}
-
-func (p Pause) Edge() inngest.Edge {
-	return inngest.Edge{
-		Outgoing: p.Outgoing,
-		Incoming: p.Incoming,
-	}
 }
 
 // Metadata must be stored for each workflow run, allowing the runner to inspect